requirements.txt
flask
requests
google-cloud-speech
google-cloud-storage
websocket-client
gunicorn